requirements.txt
fastapi
uvicorn[standard]
sqlalchemy-cockroachdb
psycopg2-binary